Many Changes
Made config.yml with some real config options and comments
Added a couple of perm nodes to plugin.yml
Added a bunch of functions to Town.java
Added TownRank.java to track town rank charges out of the config file.
Added psuedo code to Muni.java to prepare for the Town and TownRankup
classes.
The rest of the changes are just javadoc comments.