Changed storage type for Towns and Citizens
Now using a TreeSet for both towns and citizens in Muni.class.  Seems to
be working fine.
Made the compareTo work on just the names of the classes since those are
the primary keys.  A new ordering will be required for the town and
citizen list outputs to give priority to rank or officership as
applicable.
Changed some update functions in dbwrapper to use executeUpdate instead
of executeQuery, which has removed an error from the logger.