Minor changes
Added ORDER BY to getSingleCol().
Fixed checkexistence to throw an info "Value not found" error instead of
a severe when there is no record to be found in the database.
Made some public methods to add and remove from towns and citizens.
Ensuring the database is there with dbwrapper.createDB().