Bug fixes, Config additions
Code needs some deletions and commenting after that.

Town loadfromDB now loads each citizen role/type. Getting strange
results with /town info <>, some roles are the same in both test towns.
They should mostly be reporting as empty.
Towns now add citizens to the Hashmap in Muni.  Going to remove global
hashmap changes in Citizens class and reduce it to helper functions
only.
TownAdminCommand listALLcits correctly shows the content of global
hashmap and listcits (iterative on Muni.towns) works but shows the same
incorrect values as info.
Added SQL debug and use of OP to the config options.  useOP was already
implemented in the econwrapper so should be active, and changed the
dbwrapper logging decision to the new Muni.isSQLdebug().  Need to add
ingame commands to activate both debugs, and decide whether to save
these changes to config.
Created dbwrapper.getTownCits to get a list of citizens who are depty,
invitee, etc. seems to be working correctly according to the debug
output.
Created Muni.out which deals with console output automatically, location
chosen for ease of typing. ChatColor version available and plans to make
passable StringBuilder work for multiple colors in one output.
Added citizen (bool) to Citizens for database purposes.  Future: Going
to change this to an enum instead of 4 booleans, rewrite database with
types.