Changed Citizen's way of storing role
There is now a roles enum in Citizens to ensure correct spelling of role
names.  A citizen's role is tracked througt a string instead of by 5
booleans.  Most changes in this commit are a result of this change,
including constructors and database helper methods.
Added Town.getRole() to parse a player's name and state its role if any.