Town and Commands: minor edits
Made administrative commands in Town class that deal with making
citizens for the town.  Fixed many of the commands to automatically save
to the database.  More work needs to be done in this area.
Reworded (for all commands) "Not enough parameters" to "Incorrect number
of params.."
Reworked some townadmin subcommands to use the new town/citizen methods.