Updated DB structure (changed towns, added votes)
Added the votes table to allow for future addition of voting.  Structure
may or may not be final.  The table will be created at next restart
since it does not exist.
Re-added the town center location variables to the town table.  Does not
load into memory yet and table will not be loaded at next restart since
it does exist.  A drop will be needed.  This is future-proofing so that
early alpha/beta testers will not haveto deal with this DB structure
change.