Fixed town info duplcation bug
Where it says "trying to fix a bug with this" worked!  Woot woot!
wasn't clearing the data.  This method will be reworked in the future
with a database getter method that returns a list of towns instead of
one that returns a list of town names then queries individually.

Also some comments were added and code was otherwise cleaned up.