Cleanup, Bug found but not fixed
Made towns and allCitizens into TreeMap, having trouble with NPEs when
using towns.get().  Unsure of solution.  Did lots of cleanup in Muni
class.  Changed Muni.getTown() and getTownName() to use a string instead
of a player.

Many changes are due to the switch to TreeMap.

Added a new Listeners folder and MuniLoginEvent which does not fire
properly.  The code within came from Muni's class.