Moved command utils to Muni, debuged changes
Also made where you can message officers or announce to the whole town
(in the Town class).  Will add officer commands to use these functions
in the next commit.
Added saveToDB() to many of the town commands (the ones that affect
membership standing) and messages to the affected players.

Will need to remove redundant messages out of command executors.