The real Alpha 0.1 release reference
There were bugs and an issue brought up by bukkit staff that are now
resolved.  This is the code to be referenced when debugging v0.1.

Made new methods for getting a list of applicants and invitees.  Using
these new methods in the PlayerJoinEvent, which now displays relevant
town information.
Fixed the econwrapper.pay() method to work correctly with the rankup id
Fixed the getItemName() method to display the item name as capital then
lowercase (instead of all caps) and it adds an 's' on the end if its a
plural amount.