Made towns track the world
Removed the DB-autoloading constructor.  Want it to be clear higher up
that the db is loaded with an explicit call.
Added support for the townWorld string in the database and expanded the
constuctors, then added the missing field to fix the broken calls.
Re-enabling townCenter variable and need to update db calls to using
them.  Not yet finished.
Made isValid() methods for Town and Citizen, which tests for null
values.