Pseudo added for WG, voting; Towns item bal
Added pseudocode in the new class VotingManager
Working towards adding items to the town bank, added integer field to
the table, need to fix all the functions and expand dbwrapper to pulling
item balance (and town location as well).
Added commands and pseudocode to wgwrapper for what should be all the
required commands towards the second alpha.