Updated towns with democracy, item bank
Banking functions now autosave.  Item banking functions added.
Added a boolean to Towns to track democracy.  This will eventually
affect the price of establishing a town and doing rankups.  For now, it
should be saving to the db only.
DB insert and updates should be working with the new fields.

No testing has been done on any of the changes since the Alpha.