Brought reg. expansion online, some cleanup
Updated the version to v.20 but still needs work before the release.

Fixed the Town copy constructor to save this.townWorld, which was the
NPE bug with WG expansion.
Added the new WG commands to the officer help output.
Refactored in /townadmin to use "staffer" instead of "player".
Fixed up comments and minor logic in Muni.java.

To Do:
*Add an increasing cost to region expansions
*Add a flat fee per type of sub-region
*Make a shift(move) function to slightly change the location of
sub-regions
*Make the primary region automatically delete when the mayor deletes the
town
*Add a table to track the sub-regions so they can be deleted as well
*Make the number of sub-regions limited by the bounds set in the config
file.