Some work on voting #10
Added some pseudocode to dbwrapper for voting.
Made a ballot class and a vote class to encapsulate the voting data.
Removed townName from the votes table.

Planning to use "for" and "against" as the answers for binary votes.
This will allow Muni to track both types of votes in one database table